+++ resources/js/components/LinkForm.vue
@@ -3,39 +3,39 @@
         <div class="card-body">
             <div class="form-group">
                 <label for="url">{{ __('URL') }}</label>
-                <input type="text" class="form-control" :class="{'is-invalid': hasFormError('url')}" ref="url" id="url" v-model="form.url">
+                <input type="text" class="form-control" :class="{'is-invalid': hasFormError('url')}" ref="url" id="url" v-model="form.url" dusk="link-form-url">
                 <span class="invalid-feedback" v-if="hasFormError('url')">{{ firstFormError('url') }}</span>
-                <small class="form-text text-muted" v-if="parsing">{{ __('Retrieving URL informations...') }}</small>
+                <small class="form-text text-muted" v-if="parsing" dusk="link-form-parsing-message">{{ __('Retrieving URL informations...') }}</small>
             </div>
 
             <div class="form-group">
                 <label for="title">{{ __('Title') }}</label>
-                <input type="text" class="form-control" :class="{'is-invalid': hasFormError('title')}" id="title" v-model="form.title" :disabled="loading">
+                <input type="text" class="form-control" :class="{'is-invalid': hasFormError('title')}" id="title" v-model="form.title" :disabled="loading" dusk="link-form-title">
                 <span class="invalid-feedback" v-if="hasFormError('title')">{{ firstFormError('title') }}</span>
             </div>
 
             <div class="form-group">
                 <label for="content">{{ __('Content') }}</label>
-                <textarea id="content" class="form-control" :class="{'is-invalid': hasFormError('content')}" v-model="form.content" :disabled="loading"></textarea>
+                <textarea id="content" class="form-control" :class="{'is-invalid': hasFormError('content')}" v-model="form.content" :disabled="loading" dusk="link-form-content"></textarea>
                 <span class="invalid-feedback" v-if="hasFormError('content')">{{ firstFormError('content') }}</span>
             </div>
 
             <div class="form-group">
                 <div class="custom-control custom-switch">
                     <input type="checkbox" class="custom-control-input" id="is_private" v-model="form.is_private" :disabled="loading">
-                    <label class="custom-control-label" for="is_private">{{ __('Private link?') }}</label>
+                    <label class="custom-control-label" for="is_private" dusk="link-form-private">{{ __('Private link?') }}</label>
                 </div>
             </div>
 
             <div class="form-group">
                 <label>{{ __('Tags') }}</label>
-                <tags v-model="form.tags"></tags>
+                <tags v-model="form.tags" dusk="link-form-tags"></tags>
             </div>
         </div>
 
         <div class="card-footer d-flex justify-content-between">
             <div>
-                <button class="btn btn-primary" @click.prevent="submit" :disabled="loading">
+                <button class="btn btn-primary" @click.prevent="submit" :disabled="loading" dusk="link-form-save">
                     <span class="spinner-grow spinner-grow-sm" role="status" aria-hidden="true" v-if="loading"></span>
                     {{ __('Save') }}
                 </button>
@@ -120,8 +120,9 @@ export default {
                 this.loading = false;
                 this.parsing = false;
 
+                this.setFormError(error);
                 this.setHttpError(error);
-                this.toastHttpError(__("Unable to parse link"));
+                this.toastHttpError(this.__("Unable to parse link"));
             });
         },